client/consensus: initial tests - konsensus
